Added BL Manager and BL Tests, Fixed problem with stored procedure

+++ ART.Gravity/ART.Gravity.PL/spCalcForceResult.cs
@@ -8,6 +8,6 @@ namespace ART.Gravity.PL
 {
     public class spCalcForceResult
     {
-        public float Force { get; set; }
+        public double Force { get; set; }
     }
 }
